requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
python-dotenv
schedule
fake-useragent
plotly
pandas